refacto: move time function to timeUtils
